mago-mi refactoring
