+++ MagoMI/mago-mi/source/cmdline.h
@@ -16,6 +16,7 @@ struct ExecutableInfo {
 	std::wstring dir;
 	std::wstring logFile;
 	std::wstring logLevel;
+	std::wstring tty;
 	bool verbose;
 	bool miMode;
 	bool silent;
@@ -27,6 +28,7 @@ struct ExecutableInfo {
 	void setExecutable(std::wstring exe);
 	void setDir(std::wstring directory);
 	void addArg(std::wstring param);
+	void setTty(std::wstring tty);
 	void dumpParams();
 
 	bool hasExecutableSpecified();
